Remove remote content code duplication
