+++ apps/messages-app/src/adapters/outbound/rc-confguration/rc-configuration.ts
@@ -1,56 +1,19 @@
 import {
-  FiscalCodeSchema,
   GenericError,
   NotFoundError,
   TooManyRequestsError,
 } from "@pagopa/hexagonal-core";
+import {
+  type RcConfigurationResponse,
+  RcConfigurationResponseSchema,
+} from "io-messages-common/adapters/remote-content";
 import { Result, ResultAsync, err, ok } from "neverthrow";
-import z from "zod";
 
 import {
   RCConfiguration,
   RCConfigurationRepository,
+  rcConfigurationSchema,
 } from "../../../application/ports/rc-configuration.js";
-
-const rcClientCertSchema = z.object({
-  clientCert: z.string().min(1),
-  clientKey: z.string().min(1),
-  serverCa: z.string().min(1),
-});
-
-const rcAuthenticationConfigSchema = z.object({
-  cert: rcClientCertSchema.optional(),
-  headerKeyName: z.string().min(1),
-  key: z.string().min(1),
-  type: z.string().min(1),
-});
-
-const rcEnvironmentConfigSchema = z.object({
-  baseUrl: z.string().min(1),
-  detailsAuthentication: rcAuthenticationConfigSchema,
-});
-
-const rcTestEnvironmentConfigSchema = rcEnvironmentConfigSchema.extend({
-  testUsers: z.array(FiscalCodeSchema),
-});
-
-// rcConfigurationApiResponseSchema represents the remote-content configuration
-// obtained as response from the rc-app micro-service.
-const rcConfigurationApiResponseSchema = z.object({
-  configurationId: z.ulid(),
-  description: z.string().min(1),
-  disableLollipopFor: z.array(FiscalCodeSchema),
-  hasPrecondition: z.enum(["ALWAYS", "ONCE", "NEVER"]),
-  id: z.string().min(1),
-  isLollipopEnabled: z.boolean(),
-  name: z.string().min(1),
-  prodEnvironment: rcEnvironmentConfigSchema.optional(),
-  testEnvironment: rcTestEnvironmentConfigSchema.optional(),
-  userId: z.string().min(1),
-});
-type RCConfigurationApiResponse = z.infer<
-  typeof rcConfigurationApiResponseSchema
->;
 
 export class RCConfigurationHttpClientAdapter
   implements RCConfigurationRepository
@@ -62,20 +25,28 @@ export class RCConfigurationHttpClientAdapter
   }
 
   private toDomainRCConfiguration(
-    rcConfigurationApiResponse: RCConfigurationApiResponse,
-  ): RCConfiguration {
-    return {
-      configurationId: rcConfigurationApiResponse.configurationId,
-      description: rcConfigurationApiResponse.description,
-      disableLollipopFor: rcConfigurationApiResponse.disableLollipopFor,
-      hasPrecondition: rcConfigurationApiResponse.hasPrecondition,
-      id: rcConfigurationApiResponse.id,
-      isLollipopEnabled: rcConfigurationApiResponse.isLollipopEnabled,
-      name: rcConfigurationApiResponse.name,
-      prodEnvironment: rcConfigurationApiResponse.prodEnvironment,
-      testEnvironment: rcConfigurationApiResponse.testEnvironment,
-      userId: rcConfigurationApiResponse.userId,
-    };
+    response: RcConfigurationResponse,
+  ): Result<RCConfiguration, GenericError> {
+    const parsedResult = rcConfigurationSchema.safeParse({
+      configurationId: response.configurationId,
+      description: response.description,
+      disableLollipopFor: response.disableLollipopFor,
+      hasPrecondition: response.hasPrecondition,
+      id: response.id,
+      isLollipopEnabled: response.isLollipopEnabled,
+      name: response.name,
+      prodEnvironment: response.prodEnvironment,
+      testEnvironment: response.testEnvironment,
+      userId: response.userId,
+    });
+
+    return parsedResult.success
+      ? ok(parsedResult.data)
+      : err(
+          new GenericError(
+            `malformed remote content configuration returned by the rc-app: ${parsedResult.error.message}`,
+          ),
+        );
   }
 
   async getRemoteContentConfiguration(
@@ -100,7 +71,7 @@ export class RCConfigurationHttpClientAdapter
         return err(jsonResponse.error);
       }
 
-      const parsedResult = rcConfigurationApiResponseSchema.safeParse(
+      const parsedResult = RcConfigurationResponseSchema.safeParse(
         jsonResponse.value,
       );
 
@@ -111,7 +82,7 @@ export class RCConfigurationHttpClientAdapter
           ),
         );
 
-      return ok(this.toDomainRCConfiguration(parsedResult.data));
+      return this.toDomainRCConfiguration(parsedResult.data);
     }
 
     switch (response.value.status) {
